requirements.txt
streamlit==1.29.0
sqlalchemy==2.0.25
Pillow==10.2.0
google-generativeai==0.3.2
groq==0.5.0
python-dotenv==1.0.0
requests==2.31.0
mysql-connector-python
tavily-python==0.2.6
